Move the common class
