+++ forge/src/gameTest/java/com/yogpc/qp/forge/data/TestEnvironmentGeneratorForge.java
@@ -2,6 +2,7 @@ package com.yogpc.qp.forge.data;
 
 import com.mojang.serialization.Codec;
 import com.yogpc.qp.QuarryPlus;
+import com.yogpc.qp.data.AbstractTestGenerator;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.Registries;
