+++ forge/src/gameTest/java/com/yogpc/qp/forge/gametest/LoadTest.java
@@ -3,8 +3,8 @@ package com.yogpc.qp.forge.gametest;
 import com.kotori316.testutil.common.TestFunction;
 import com.yogpc.qp.PlatformAccess;
 import com.yogpc.qp.QuarryPlus;
+import com.yogpc.qp.data.GatherGameTest;
 import com.yogpc.qp.forge.PlatformAccessForge;
-import com.yogpc.qp.forge.data.GatherGameTest;
 import com.yogpc.qp.gametest.GameTestFunctions;
 import net.minecraft.gametest.framework.GameTestHelper;
 import net.minecraft.network.chat.Component;
